+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 - This is not a V14 update. An update for V14 is still a work in progress.
 ## New Modern Content:
 - Unarmed Strike (Monk, modern)
+- Animate Objects
 ## New Legacy Content:
 - Martial Arts
 - Slow Fall
-- Animate Objects
+- Animate Objects (Thanks BinaryScrub)
 ## Bug Fixes:
 - Iron Chin
 - Fisticuffs